+++ src/services/AdendaService.ts
@@ -3,16 +3,15 @@ import { adenda, pessoa, veiculo } from "@prisma/client";
 import AdendaRepository from "../repositories/mysql/AdendaRepository";
 import { validateAdendaDates } from "../utils/helper";
 import ApoliceRepository from "../repositories/mysql/ApoliceRepository";
+import CustomError from "../utils/CustomError";
 @Service()
 class AdendaService {
  
-
     @Inject(() => AdendaRepository)
     private adenda_repo: AdendaRepository;
 
     @Inject(() => ApoliceRepository)
     private apolice_repo: ApoliceRepository;
-
 
     constructor() { }
 
@@ -22,6 +21,11 @@ class AdendaService {
 
     async getByID(id: string): Promise<adenda> {
         return this.adenda_repo.getByID(id);
+    }
+
+    async exists(id: string): Promise<boolean> {
+        const item = await this.getByID(id);
+        return item !== null && item !== undefined;
     }
 
     async getByApoliceID(id: string):Promise<adenda[]> {
@@ -48,6 +52,9 @@ class AdendaService {
 
     async remover(id: string): Promise<boolean> {
         // check if object exist
+        if(!await this.exists(id)) {
+            throw new CustomError("A adenda que deseja remover não existe!");
+        }
         return this.adenda_repo.delete(id);
     }
 
@@ -63,7 +70,7 @@ class AdendaService {
     }
 
     async adicionarSegurados(id: string, segurados:pessoa[]): Promise<pessoa[]>{
-        return this.apolice_repo.addSeguradosByAdendaID(id, segurados);
+        return this.adenda_repo.addSeguradosByAdendaID(id, segurados);
     }
 }
