+++ src/services/CoberturaService.ts
@@ -2,6 +2,7 @@ import 'reflect-metadata';
 import { Service, Inject } from "typedi";
 import CoberturaRepository from "../repositories/mysql/CoberturaRepository";
 import { cobertura } from '@prisma/client';
+import CustomError from '../utils/CustomError';
 
 // Add the business logic here
 @Service()
@@ -19,6 +20,11 @@ class CoberturaService {
         return this.repo.getByID(id);
     }
 
+    async exists(id: string): Promise<boolean> {
+        const item = await this.getByID(id);
+        return item !== null && item !== undefined;
+    }
+
     async getByApoliceTipo(id: string): Promise<cobertura>{
         return this.repo.getByID(id);
     }
@@ -33,8 +39,12 @@ class CoberturaService {
 
     async remover(id: string): Promise<boolean> {
         // check if object exist
+        if(!await this.exists(id)) {
+            throw new CustomError("A Cobertura que deseja remover não existe!");
+        }
         return this.repo.delete(id);
     }
+
 }
 
 export default CoberturaService;
