+++ src/services/PrecoCilindradaService.ts
@@ -2,6 +2,7 @@ import { Service, Inject } from "typedi";
 import PrecoCilindradaRepository from "../repositories/mysql/PrecoCilindradaRepository";
 import IGenericRepository from "../repositories/IGenericRepository";
 import { preco_cilindrada } from "@prisma/client";
+import CustomError from "../utils/CustomError";
 
 @Service()
 class PrecoCilindradaService {
@@ -18,6 +19,11 @@ class PrecoCilindradaService {
         return this.repo.getByID(id);
     }
 
+    async exists(id: string): Promise<boolean> {
+        const item = await this.getByID(id);
+        return item !== null && item !== undefined;
+    }
+
     async criar(preco_cilindrada: preco_cilindrada): Promise<preco_cilindrada> {
         return this.repo.create(preco_cilindrada);
     } 
@@ -28,6 +34,9 @@ class PrecoCilindradaService {
 
     async remover(id: string): Promise<boolean> {
         // check if object exist
+        if(!await this.exists(id)) {
+            throw new CustomError("A O preço por cilindrada que deseja remover não existe!");
+        }
         return this.repo.delete(id);
     }
 }
